refactor: moved all finetune logic to separate package
