+++ finetune/__init__.py
@@ -1,0 +1,6 @@
+"""
+This package serves to be only used to finetune the model
+"""
+from .finetune import FineTuner, CustomDataCollator
+
+__all__ = ['FineTuner', 'CustomDataCollator']
